Minor dir restructuring for gcloud
