+++ cmd/main.go
@@ -1,0 +1,24 @@
+// This file is purely for local development
+
+package main
+
+import (
+	"log"
+	"os"
+
+	"github.com/GoogleCloudPlatform/functions-framework-go/funcframework"
+	"playercount"
+)
+
+func main() {
+	funcframework.RegisterHTTPFunction("/", playercount.HelloWorld)
+	// Use PORT environment variable, or default to 8080.
+	port := "8080"
+	if envPort := os.Getenv("PORT"); envPort != "" {
+		port = envPort
+	}
+
+	if err := funcframework.Start(port); err != nil {
+		log.Fatalf("funcframework.Start: %v\n", err)
+	}
+}
