+++ go.mod
@@ -3,6 +3,7 @@ module playercount
 go 1.13
 
 require (
+	github.com/GoogleCloudPlatform/functions-framework-go v1.0.1
 	github.com/PuerkitoBio/goquery v1.5.1
 	github.com/cheggaaa/pb/v3 v3.0.4
 	github.com/joho/godotenv v1.3.0
